UIT-21: Rename AuthenticationProvider to AuthenticationContext
